Now tries to match the right constructor given the type mappings
It now finds the "overlap" between the object's set of constructors and the mapped types. It looks for the greatest overlap and selects that constructor. It doesn't do anything about ambiguous matches though (if it matches more than one constructor). This can occur if a type has two constructors that take the same type that's been mapped.