Now throws AmbiguousConstructorException if multiple constructors match
This occurs when building an object when the type map matches multiple constructors. For example, if you map a string, and the object has two constructors that both take strings, it doesn't know which one to use, since the mapping isn't specific enough. In this case, a new exception is thrown to let the developer know that the type map needs more mappings.

Of course, this also adds the new exception type being thrown.

Had to update an old test (the Uri one), since its constructor now fell into this ambiguous scenario.