Now indicates all found constructors when throwing AmbiguousConstructor
This is a message for the developer to let them know what constructors were matched with their given type map.

Uses a static "Resources" class for strings. Probably won't be translated, but it still extracts the strings out of the logic class into a separate specialized class for holding them.