Refactoring the type map object into its own domain object
Now it lives in a TypeMapEntry, which houses the instance still, but also makes room for extra data. So far we'll use the "has been mapped" bool so that when matching constructor parameters we only use one instance once. This prevents constructors that take the same type multiple times from using the same INSTANCE multiple times.