Now asks "For<T>" a builder, instead of having to call Create()
This re-works the API, changing it from an ObjectBuilder-centric model (specifying the type parameter on that guy), so calling a For<T> method on ObjectBuilder, who then returns a different object to drill into. The use feels a little nicer, since the type parameter is more closely-related to what you're asking for.

This means ObjectBuilder (non-generic) acts as a factory to ObjectBuilder<T>, who now does all the heavy lifting. This separates concerns and keeps the API profile slim.