+Blank AcceptanceTests project
This will be the project that are basically integration tests, end-to-end, assuring that we want user stories to be true through specific API interactions.

With the change to "MapInterfacesTo" only being callable once (through API design, not through state tracking), we should verify that the cases around this are true. Seems like a good time to start this brand of testing.