+ParameterName struct
This is used only for the API usability, not for any functionality. Soon we'll have an overload that allows With<>() to take a Func that PROVIDES a value allowing the caller to name the variable, like this:

   With( parameter => 3 )

The "parameter" variable is of type ParameterName. When you mouse over this variable in the lambda, it literally says "ParameterName" to give a hint that it's only a parameter. Really, it's unused. The important part is that you return a value that matches whatever parameter you're configuring.

It's also a struct so we don't make objects needlessly for garbage collection.